move configuration loading logic to a service class and implement env variables
