+++ docker-compose.yaml
@@ -8,6 +8,8 @@ services:
     container_name: api-gateway
     ports:
       - "8080:8080"
+    environment:
+      ONTOPORTAL_APIKEY: "put here APIKEY"
     restart: always
     networks:
       - api-gateway
